Move most privileged actions to image service.
This will allow us to start the electron UI in the startup menu automatically.

There will still be a UAC prompt on the first boot due to `mkcert` generation (this could be done in C# in the service wrapper without too much hassle) for installations that don't enable the future "start on boot" that will place the electron UI in startup and enable the nginx and image server services on boot. However, with "start on boot" enabled, the electron side will have no need to ever request elevated permissions.